Added top_annotation to produce_heatmap function
It is requiered to add top_annotation to the heatmaps when working with
partitions to make sure the colors between the top_annotation and the
partition heatmap are the same.

The param partitions was thus added to the produce_heatmap function.

Also, to make sure the color remain the same between the top_annotation
and the partition heatmap, a seed parameter was added. The users will
just have to make to to use the same value when calling produce_heatmap
and preprare_heatmap_list.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>